requirements-api.txt
# ACSO API Gateway Requirements

# Core FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication and Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==41.0.7
bcrypt==4.1.2

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0  # PostgreSQL async driver
psycopg2-binary==2.9.9  # PostgreSQL sync driver

# Redis and Caching
redis==5.0.1
aioredis==2.0.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Data Validation and Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# WebSocket Support
websockets==12.0

# Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0

# AWS SDK (for future integration)
boto3==1.34.0
botocore==1.34.0

# Optional: Database migrations
yoyo-migrations==8.2.0